InsurePredict AI/backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
numpy
pandas
scikit-learn
